requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.5
Pillow==10.4.0
gunicorn==23.0.0
